+++ tests/private/users.py
@@ -85,7 +85,8 @@ class UsersTests(ApiTestBase):
         self.assertIsNotNone(results.get('geo_media'))
 
     def test_search_users(self):
-        results = self.api.search_users('maruhanamogu')
+        rank_token = self.api.generate_uuid()
+        results = self.api.search_users('maruhanamogu', rank_token)
         self.assertEqual(results.get('status'), 'ok')
 
     def test_check_username(self):